DXSDE: Dynamically generated XML editor from an XSD
This project will attempt to produce a dynamically generated XML editor from an
XSD, the idea being that it can be used as a component in other applications
for Rapid Application Development (and relieve the tedium of making multiple,
slightly different editing forms for each object).

The primary project is in C#, however a Perl version is in the pipeline.